Restructure database
Add user_id in the currencies and the types.
Undo of the deleted code from the types (creation and delete)
Add the subtraction field to the type. This allow the user to create new
types and check if is to subtract in the balance or not.
Improve some UI stuff